🛠 Minimal (without restriction) cards selection added #28 If no tags is provided all the card, sorted by date are searched
