--- frontend/mobile_app_react_native/local_storage/insertData.ts
@@ -1,16 +1,0 @@
-import {SQLiteDatabase } from 'react-native-sqlite-storage';
-import { tableName } from '../constants/local_storage';
-
-interface dataItem {
-    tag : String,
-    card : String
-}
-const insertData = async (db: SQLiteDatabase, dataItems: dataItem) => {  
-  const insertQuery =
-      `INSERT OR REPLACE INTO ${tableName} (value)
-        values 
-         ('${dataItems.tag}');` ;
-    return db.executeSql(insertQuery);
-};
-
-export default insertData;
