+++ frontend/mobile_app_react_native/local_storage/main.ts
@@ -1,5 +1,4 @@
 import createTable from "./createTable";
 import getDBConnection from "./getDBConnection";
-import insertData from "./insertData";
 
-export {createTable, getDBConnection, insertData}; 
+export {createTable, getDBConnection};
